MolecularGraph moved to rotamer
